FE-Refactor: web/src 폴더 제거
