+++ apps/web/app/admin/members/_components/ResponsiveMembersPage.tsx
@@ -3,7 +3,7 @@
 import { PAGE_NAME } from "@ui/src/utils/constants/pageNames";
 import ErrorResetBoundary from "@/components/common/ErrorResetBoundary";
 import ErrorFallback from "@/components/common/Fallback";
-import { useRedirectOnMobile } from "@/src/hooks/useRediectOnMobile";
+import { useRedirectOnMobile } from "@/app/_hooks/useRediectOnMobile";
 import MembersContent from "./MembersContent";
 
 export default function ResponsiveMembersPage(): JSX.Element {
