+++ apps/web/app/admin/teams/_components/ResponsiveTeamsPage.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense } from "react";
 import { PAGE_NAME } from "@ui/src/utils/constants/pageNames";
-import { useRedirectOnMobile } from "@/src/hooks/useRediectOnMobile";
+import { useRedirectOnMobile } from "@/app/_hooks/useRediectOnMobile";
 import ErrorResetBoundary from "@/components/common/ErrorResetBoundary";
 import ErrorFallback from "@/components/common/Fallback";
 import TeamListHeader from "./TeamListHeader";
