+++ apps/web/app/context/SeatContext.tsx
@@ -4,7 +4,7 @@ import { createContext, useContext, useState, type ReactNode, useMemo } from "re
 import { useQuery } from "@tanstack/react-query";
 import { type IReservation } from "@repo/types";
 import { getUserReservations } from "@/api/reservation";
-import { useAuthStore } from "../stores/useAuthStore";
+import { useAuthStore } from "../store/useAuthStore";
 
 interface SeatContextType {
   checkedSeat: string | null;
