+++ apps/web/app/utils/seats.ts
